[3.6] bpo-34272: Move argument parsing tests from test_capi to test_getargs2. (GH-8567). (cherry picked from commit 8f7bb100d0fa7fb2714f3953b5b627878277c7c6)
